Bunch of fixes and improvements
Been working on this for a while now. Fixed a lot of the missing pieces that were causing issues. The codegen was pretty broken - match statements weren't working at all, and the iterator stuff for for-loops was just stubbed out.

Got the semantic analysis working better too. Was having problems with type checking on some of the more complex expressions. The pattern matching in particular needed a lot of work.

Also spent time on the WASM backend which had some gaps. Array operations and struct creation weren't implemented properly.

The transaction execution was another pain point - it wasn't actually calling into the VM to run contracts, just doing basic transfers. Now it properly integrates with the AVM.

Still not perfect but much more functional now. Should be able to compile and run most programs without hitting unimplemented stubs.